week6/django_rest_example/requirements.txt
Django==1.10.3
djangorestframework==3.5.3
nose==1.3.7
psycopg2==2.6.2
requests==2.12.1
